+++ APIBlox.AspNetCore/Contracts/IProjectedQuery.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace APIBlox.AspNetCore.Contracts
+{
+    /// <summary>
+    ///     Interface IProjectedQuery
+    /// </summary>
+    public interface IProjectedQuery
+    {
+        
+        /// <summary>
+        ///     Gets or sets the select.
+        /// </summary>
+        /// <value>The select.</value>
+        string Select { get; set; }
+    }
+}