+++ APIBlox.AspNetCore/Extensions/QueryableExtensions.cs
@@ -1,10 +1,8 @@
 ﻿#region -    Using Statements    -
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Dynamic.Core;
-using System.Threading.Tasks;
 using APIBlox.AspNetCore.Contracts;
 using APIBlox.AspNetCore.Types;
 using APIBlox.NetCore.Extensions;
@@ -21,29 +19,39 @@ namespace APIBlox.AspNetCore.Extensions
     public static class QueryableExtensions
     {
         /// <summary>
-        ///     Builds results from a queryable as an asynchronous operation, that assumes client side paging.
+        ///     Builds a single result query from a queryable.
         /// </summary>
         /// <typeparam name="TIn">The type of the t in.</typeparam>
         /// <typeparam name="TOut">The type of the t out.</typeparam>
         /// <param name="qry">The qry.</param>
-        /// <param name="projection">Select projection</param>
-        /// <param name="serializeSettings">Map serialize settings</param>
-        /// <returns>Task&lt;List&lt;TOut&gt;&gt;.</returns>
-        public static Task<List<TOut>> ResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+        /// <param name="request">The request.</param>
+        /// <param name="projection">The projection.</param>
+        /// <param name="serializeSettings">The serialize settings.</param>
+        /// <returns>Task&lt;TOut&gt;.</returns>
+        public static IQueryable ProjectedResultQuery<TIn, TOut>(this IQueryable<TIn> qry,
+            IProjectedQuery request,
             Func<TIn, TOut> projection = null,
             JsonSerializerSettings serializeSettings = null
         )
             where TOut : new()
         {
-            var result = qry.BuildOutputQueryable(projection, serializeSettings);
+            var tmp = qry.BuildOutputQueryable(projection, serializeSettings);
+
+            if (request.Select.IsEmptyNullOrWhiteSpace())
+            {
+                Console.WriteLine(tmp.Expression.ToString());
+                return tmp;
+            }
+
+            var result = tmp.BuildUserProjectionQueryable(request);
 
             Console.WriteLine(result.Expression.ToString());
 
-            return result.ToDynamicListAsync<TOut>();
+            return result;
         }
 
         /// <summary>
-        ///     Builds a ordered result from a queryable as an asynchronous operation, that assumes client side paging.
+        ///     Builds a ordered result query from a queryable, that assumes client side paging.
         /// </summary>
         /// <typeparam name="TIn">The type of the t in.</typeparam>
         /// <typeparam name="TOut">The type of the t out.</typeparam>
@@ -52,7 +60,7 @@ namespace APIBlox.AspNetCore.Extensions
         /// <param name="projection">Select projection</param>
         /// <param name="serializeSettings">Map serialize settings</param>
         /// <returns>Task&lt;List&lt;TOut&gt;&gt;.</returns>
-        public static Task<List<TOut>> OrderedResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+        public static IQueryable OrderedResultsQuery<TIn, TOut>(this IQueryable<TIn> qry,
             IOrderedQuery request,
             Func<TIn, TOut> projection = null,
             JsonSerializerSettings serializeSettings = null
@@ -64,11 +72,11 @@ namespace APIBlox.AspNetCore.Extensions
 
             Console.WriteLine(result.Expression.ToString());
 
-            return result.ToDynamicListAsync<TOut>();
+            return result;
         }
 
         /// <summary>
-        ///     Builds a ordered result from a filtered queryable as an asynchronous operation, that assumes client side paging.
+        ///     Builds a ordered result query from a filtered queryable, that assumes client side paging.
         /// </summary>
         /// <typeparam name="TIn">The type of the t in.</typeparam>
         /// <typeparam name="TOut">The type of the t out.</typeparam>
@@ -77,7 +85,7 @@ namespace APIBlox.AspNetCore.Extensions
         /// <param name="projection">Select projection</param>
         /// <param name="serializeSettings">Map serialize settings</param>
         /// <returns>Task&lt;List&lt;dynamic&gt;&gt;.</returns>
-        public static Task<List<dynamic>> FilteredResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+        public static IQueryable FilteredResultsQuery<TIn, TOut>(this IQueryable<TIn> qry,
             IFilteredQuery request,
             Func<TIn, TOut> projection = null,
             JsonSerializerSettings serializeSettings = null
@@ -92,18 +100,18 @@ namespace APIBlox.AspNetCore.Extensions
             if (request.Select.IsEmptyNullOrWhiteSpace())
             {
                 Console.WriteLine(tmp.Expression.ToString());
-                return tmp.ToDynamicListAsync();
+                return tmp;
             }
 
             var result = tmp.BuildUserProjectionQueryable(request);
 
             Console.WriteLine(result.Expression.ToString());
 
-            return result.ToDynamicListAsync();
+            return result;
         }
 
         /// <summary>
-        ///     Builds a paged result from a filtered queryable as an asynchronous operation, where server side paging is enforced.
+        ///     Builds a paged result query from a filtered queryable, where server side paging is enforced.
         /// </summary>
         /// <typeparam name="TIn">The type of the t in.</typeparam>
         /// <typeparam name="TOut">The type of the t out.</typeparam>
@@ -113,7 +121,7 @@ namespace APIBlox.AspNetCore.Extensions
         /// <param name="projection">Select projection</param>
         /// <param name="serializeSettings">Map serialize settings</param>
         /// <returns>Task&lt;List&lt;TOut&gt;&gt;.</returns>
-        public static Task<List<dynamic>> FilteredPagedResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+        public static IQueryable FilteredPagedResultsQuery<TIn, TOut>(this IQueryable<TIn> qry,
             FilteredPaginationQuery request, int maxItemsPerResult,
             Func<TIn, TOut> projection = null,
             JsonSerializerSettings serializeSettings = null
@@ -128,14 +136,14 @@ namespace APIBlox.AspNetCore.Extensions
             if (request.Select.IsEmptyNullOrWhiteSpace())
             {
                 Console.WriteLine(tmp.Expression.ToString());
-                return tmp.ToDynamicListAsync();
+                return tmp;
             }
 
             var result = tmp.BuildUserProjectionQueryable(request);
 
             Console.WriteLine(result.Expression.ToString());
 
-            return result.ToDynamicListAsync();
+            return result;
         }
 
         /// <summary>
@@ -149,7 +157,7 @@ namespace APIBlox.AspNetCore.Extensions
         /// <param name="projection">Select projection</param>
         /// <param name="serializeSettings">Map serialize settings</param>
         /// <returns>Task&lt;List&lt;dynamic&gt;&gt;.</returns>
-        public static Task<List<TOut>> PagedResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+        public static IQueryable PagedResultsQuery<TIn, TOut>(this IQueryable<TIn> qry,
             IPaginationQuery request, int maxItemsPerResult,
             Func<TIn, TOut> projection = null,
             JsonSerializerSettings serializeSettings = null
@@ -161,8 +169,9 @@ namespace APIBlox.AspNetCore.Extensions
 
             Console.WriteLine(result.Expression.ToString());
 
-            return result.ToDynamicListAsync<TOut>();
+            return result;
         }
+
 
 
         private static IQueryable<TIn> BuildFilteredQueryable<TIn>(this IQueryable<TIn> source,
@@ -192,7 +201,7 @@ namespace APIBlox.AspNetCore.Extensions
         }
 
         private static IQueryable BuildUserProjectionQueryable<TIn>(this IQueryable<TIn> source,
-            IFilteredQuery request)
+            IProjectedQuery request)
         {
             return request.Select.IsEmptyNullOrWhiteSpace() ? source : source.Select(request.Select);
         }
@@ -220,3 +229,228 @@ namespace APIBlox.AspNetCore.Extensions
         }
     }
 }
+
+
+
+//#region -    Using Statements    -
+
+//using System;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Linq.Dynamic.Core;
+//using System.Threading.Tasks;
+//using APIBlox.AspNetCore.Contracts;
+//using APIBlox.AspNetCore.Types;
+//using APIBlox.NetCore.Extensions;
+//using APIBlox.NetCore.Types;
+//using Newtonsoft.Json;
+
+//#endregion
+
+//namespace APIBlox.AspNetCore.Extensions
+//{
+//    /// <summary>
+//    ///     Class QueryableExtensions.
+//    /// </summary>
+//    public static class QueryableExtensions
+//    {
+//        /// <summary>
+//        ///     Builds results from a queryable as an asynchronous operation, that assumes client side paging.
+//        /// </summary>
+//        /// <typeparam name="TIn">The type of the t in.</typeparam>
+//        /// <typeparam name="TOut">The type of the t out.</typeparam>
+//        /// <param name="qry">The qry.</param>
+//        /// <param name="projection">Select projection</param>
+//        /// <param name="serializeSettings">Map serialize settings</param>
+//        /// <returns>Task&lt;List&lt;TOut&gt;&gt;.</returns>
+//        public static Task<List<TOut>> ResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+//            Func<TIn, TOut> projection = null,
+//            JsonSerializerSettings serializeSettings = null
+//        )
+//            where TOut : new()
+//        {
+//            var result = qry.BuildOutputQueryable(projection, serializeSettings);
+
+//            Console.WriteLine(result.Expression.ToString());
+
+//            return result.ToDynamicListAsync<TOut>();
+//        }
+
+//        /// <summary>
+//        ///     Builds a ordered result from a queryable as an asynchronous operation, that assumes client side paging.
+//        /// </summary>
+//        /// <typeparam name="TIn">The type of the t in.</typeparam>
+//        /// <typeparam name="TOut">The type of the t out.</typeparam>
+//        /// <param name="qry">The qry.</param>
+//        /// <param name="request">The request.</param>
+//        /// <param name="projection">Select projection</param>
+//        /// <param name="serializeSettings">Map serialize settings</param>
+//        /// <returns>Task&lt;List&lt;TOut&gt;&gt;.</returns>
+//        public static Task<List<TOut>> OrderedResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+//            IOrderedQuery request,
+//            Func<TIn, TOut> projection = null,
+//            JsonSerializerSettings serializeSettings = null
+//        )
+//            where TOut : new()
+//        {
+//            var result = qry.BuildOrderedQueryable(request)
+//                .BuildOutputQueryable(projection, serializeSettings);
+
+//            Console.WriteLine(result.Expression.ToString());
+
+//            return result.ToDynamicListAsync<TOut>();
+//        }
+
+//        /// <summary>
+//        ///     Builds a ordered result from a filtered queryable as an asynchronous operation, that assumes client side paging.
+//        /// </summary>
+//        /// <typeparam name="TIn">The type of the t in.</typeparam>
+//        /// <typeparam name="TOut">The type of the t out.</typeparam>
+//        /// <param name="qry">The qry.</param>
+//        /// <param name="request">The request.</param>
+//        /// <param name="projection">Select projection</param>
+//        /// <param name="serializeSettings">Map serialize settings</param>
+//        /// <returns>Task&lt;List&lt;dynamic&gt;&gt;.</returns>
+//        public static Task<List<dynamic>> FilteredResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+//            IFilteredQuery request,
+//            Func<TIn, TOut> projection = null,
+//            JsonSerializerSettings serializeSettings = null
+//        )
+//            where TOut : new()
+//        {
+//            var tmp = qry
+//                .BuildOrderedQueryable(request)
+//                .BuildFilteredQueryable(request)
+//                .BuildOutputQueryable(projection, serializeSettings);
+
+//            if (request.Select.IsEmptyNullOrWhiteSpace())
+//            {
+//                Console.WriteLine(tmp.Expression.ToString());
+//                return tmp.ToDynamicListAsync();
+//            }
+
+//            var result = tmp.BuildUserProjectionQueryable(request);
+
+//            Console.WriteLine(result.Expression.ToString());
+
+//            return result.ToDynamicListAsync();
+//        }
+
+//        /// <summary>
+//        ///     Builds a paged result from a filtered queryable as an asynchronous operation, where server side paging is enforced.
+//        /// </summary>
+//        /// <typeparam name="TIn">The type of the t in.</typeparam>
+//        /// <typeparam name="TOut">The type of the t out.</typeparam>
+//        /// <param name="qry">The qry.</param>
+//        /// <param name="request">The request.</param>
+//        /// <param name="maxItemsPerResult">max items that should override the queries top.</param>
+//        /// <param name="projection">Select projection</param>
+//        /// <param name="serializeSettings">Map serialize settings</param>
+//        /// <returns>Task&lt;List&lt;TOut&gt;&gt;.</returns>
+//        public static Task<List<dynamic>> FilteredPagedResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+//            FilteredPaginationQuery request, int maxItemsPerResult,
+//            Func<TIn, TOut> projection = null,
+//            JsonSerializerSettings serializeSettings = null
+//        )
+//            where TOut : new()
+//        {
+//            var tmp = qry.BuildPagedQueryable(request, maxItemsPerResult)
+//                .BuildOrderedQueryable(request)
+//                .BuildFilteredQueryable(request)
+//                .BuildOutputQueryable(projection, serializeSettings);
+
+//            if (request.Select.IsEmptyNullOrWhiteSpace())
+//            {
+//                Console.WriteLine(tmp.Expression.ToString());
+//                return tmp.ToDynamicListAsync();
+//            }
+
+//            var result = tmp.BuildUserProjectionQueryable(request);
+
+//            Console.WriteLine(result.Expression.ToString());
+
+//            return result.ToDynamicListAsync();
+//        }
+
+//        /// <summary>
+//        ///     Builds a paged result from a queryable as an asynchronous operation, where server side paging is enforced.
+//        /// </summary>
+//        /// <typeparam name="TIn">The type of the t in.</typeparam>
+//        /// <typeparam name="TOut">The type of the t out.</typeparam>
+//        /// <param name="qry">The qry.</param>
+//        /// <param name="request">The request.</param>
+//        /// <param name="maxItemsPerResult">max items that should override the queries top.</param>
+//        /// <param name="projection">Select projection</param>
+//        /// <param name="serializeSettings">Map serialize settings</param>
+//        /// <returns>Task&lt;List&lt;dynamic&gt;&gt;.</returns>
+//        public static Task<List<TOut>> PagedResultsAsync<TIn, TOut>(this IQueryable<TIn> qry,
+//            IPaginationQuery request, int maxItemsPerResult,
+//            Func<TIn, TOut> projection = null,
+//            JsonSerializerSettings serializeSettings = null
+//        )
+//            where TOut : new()
+//        {
+//            var result = qry.BuildPagedQueryable(request, maxItemsPerResult)
+//                .BuildOutputQueryable(projection, serializeSettings);
+
+//            Console.WriteLine(result.Expression.ToString());
+
+//            return result.ToDynamicListAsync<TOut>();
+//        }
+
+
+//        private static IQueryable<TIn> BuildFilteredQueryable<TIn>(this IQueryable<TIn> source,
+//            IFilteredQuery request
+//        )
+//        {
+//            return request.Filter.IsEmptyNullOrWhiteSpace() ? source : source.Where(request.Filter);
+//        }
+
+//        private static IQueryable<TIn> BuildOrderedQueryable<TIn>(this IQueryable<TIn> source,
+//            IOrderedQuery request
+//        )
+//        {
+//            return request.OrderBy.IsEmptyNullOrWhiteSpace() ? source : source.OrderBy(request.OrderBy);
+//        }
+
+//        private static IQueryable<TIn> BuildPagedQueryable<TIn>(this IQueryable<TIn> source,
+//            IPaginationQuery request, int maxItemsPerResult
+//        )
+//        {
+//            var max = request.Top.HasValue && request.Top.Value <= maxItemsPerResult ? request.Top.Value : maxItemsPerResult;
+
+//            source = request.Skip is null ? source : source.Skip(request.Skip.Value);
+//            source = source.Take(max);
+
+//            return source;
+//        }
+
+//        private static IQueryable BuildUserProjectionQueryable<TIn>(this IQueryable<TIn> source,
+//            IFilteredQuery request)
+//        {
+//            return request.Select.IsEmptyNullOrWhiteSpace() ? source : source.Select(request.Select);
+//        }
+
+//        private static IQueryable<TOut> BuildOutputQueryable<TIn, TOut>(this IQueryable<TIn> qry,
+//            Func<TIn, TOut> projection,
+//            JsonSerializerSettings serializeSettings
+//        )
+//            where TOut : new()
+//        {
+//            var ss = serializeSettings;
+
+//            var source = (projection is null
+//                    ? qry.Select(m => m.MapTo(default(TOut), ss))
+//                    : qry.Select(m => Projection(projection, m)))
+//                .AsQueryable();
+
+//            return source;
+//        }
+
+//        private static TOut Projection<TIn, TOut>(Func<TIn, TOut> projection, TIn m)
+//            where TOut : new()
+//        {
+//            return projection(m);
+//        }
+//    }
+//}