+++ releaseNotes.md
@@ -1,3 +1,12 @@
+_August 13th, 2019_  **v2.1.15**
+- APIBlox.AspNetCore.Types - Added IProjectedQuery and ProjectedQuery, and changed the inheritancce around a bit.  This was done to allow a GET for example to retrieve a single result, with could be projected, previously the ORDERBY and FILTER mechanisms made no sense in a single result response.
+
+Thanks,  
+_Slacquer_  
+
+<br>
+<br>
+
 _August 11th, 2019_  **v2.1.14**
 - APIBlox.NetCore.Decorators.Commands.RetryCommandHandlerDecorator, WoW! did not break if successful.
 - APIBlox.AspNetCore.TypesHandlerResponseExtensions, changed comments and method names to reflect what is in _APIBlox.AspNetCore.Enums.CommonStatusCodes_.